issue #1088 move hbase code to correct folder
