+++ examples/streaming/kafka-hbase-pipeline/README.md
